+++ resteasy-jaxrs/src/main/java/org/jboss/resteasy/spi/metadata/MethodParameter.java
@@ -23,7 +23,7 @@ public class MethodParameter extends Parameter
    @Override
    public AccessibleObject getAccessibleObject()
    {
-      return locator.method;
+      return locator.getMethod();
    }
 
    public Annotation[] getAnnotations()
