+++ resteasy-jaxrs/src/main/java/org/jboss/resteasy/spi/metadata/ResourceMethod.java
@@ -1,9 +1,8 @@
 package org.jboss.resteasy.spi.metadata;
 
 import javax.ws.rs.core.MediaType;
-import java.lang.annotation.Annotation;
+
 import java.lang.reflect.Method;
-import java.lang.reflect.Type;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -11,7 +10,7 @@ import java.util.Set;
  * @author <a href="mailto:bill@burkecentral.com">Bill Burke</a>
  * @version $Revision: 1 $
  */
-public class ResourceMethod extends ResourceLocator
+public class ResourceMethod extends DefaultResourceLocator
 {
    private static final MediaType[] empty = {};
    protected Set<String> httpMethods = new HashSet<String>();
